+++ routelandia/src/main/java/edu/pdx/its/portal/routelandia/DownloadListofHighway.java
@@ -25,6 +25,7 @@ import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
+import edu.pdx.its.portal.routelandia.entities.*;
 
 /**
  * Created by loc on 2/6/15.
