+++ routelandia/src/main/java/edu/pdx/its/portal/routelandia/MapsActivity.java
@@ -41,6 +41,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 
+import edu.pdx.its.portal.routelandia.entities.*;
+
 public class MapsActivity extends FragmentActivity {
 
     private GoogleMap mMap; // Might be null if Google Play services APK is not available.
