+++ routelandia/src/main/java/edu/pdx/its/portal/routelandia/ParserTask.java
@@ -26,6 +26,8 @@ import org.json.JSONArray;
 import java.util.ArrayList;
 import java.util.List;
 
+import edu.pdx.its.portal.routelandia.entities.*;
+
 /**
  * Created by loc on 1/24/15.
  */
