+++ routelandia/src/main/java/edu/pdx/its/portal/routelandia/entities/ApiEntity.java
@@ -1,0 +1,14 @@
+package edu.pdx.its.portal.routelandia.entities;
+
+/**
+ * This is the root class for all entities in the project. It provides the common abilities,
+ * such as reaching out to the routelandia API.
+ *
+ * Created by joshproehl on 3/2/15.
+ */
+public class ApiEntity {
+    // Define the root URL we'll concatenate individual resources URL's to
+    public static final String API_ROOT = "http://capstoneaa.cs.pdx.edu/api/";
+
+
+}
